Start defining how the Vue stuff is structured and fiddles around that.
