--- src/components/by-robots-icon/index.js
@@ -1,14 +1,0 @@
-/**
- * External
- */
-import Vue from 'vue';
-
-/**
- * Internal
- */
-import byRobotsIcon from './template.vue';
-import './style.scss';
-
-Vue.component('by-robots-icon', byRobotsIcon);
-
-export default byRobotsIcon;
