--- src/components/by-robots-icon/style.scss
@@ -1,1 +1,0 @@
-@import '../../sass/organise';
