--- src/components/by-robots-icon/template.vue
@@ -1,12 +1,0 @@
-<template>
-  <a href="#">
-    <img 
-      src="/dist/by-robots.svg" 
-      alt="By Robots" 
-      class="by-robots-icon" >
-  </a>
-</template>
-
-<script>
-    export default {};
-</script>
