--- src/components/wysiwyg/index.js
@@ -1,16 +1,0 @@
-/**
- * External
- */
-import Vue from 'vue';
-
-/**
- * Internal
- */
-import wysiwyg from './template.vue';
-import './style.scss';
-
-/**
- * Declare the component and export it
- */
-Vue.component('wysiwyg', wysiwyg);
-export default wysiwyg;
