--- src/components/wysiwyg/style.scss
@@ -1,6 +1,0 @@
-@import '../../sass/organise';
-@import '~simplemde/dist/simplemde.min.css';
-
-.wysiwyg-editor {
-    //
-}
