+++ src/instances/login/index.js
@@ -6,9 +6,7 @@ import Vue from 'vue';
 /**
  * Internal
  */
-import footerSignoff from '../../components/footer-signoff';
 import login from './login.vue';
-import loginForm from '../../components/login-form';
 import instance from '../../mixins/instance.js';
 
 import './style.scss';
