+++ src/instances/post-create/index.js
@@ -8,14 +8,12 @@ import Vue from 'vue';
  */
 import postCreate from './post-create.vue';
 import instance from '../../mixins/instance.js';
-import wysiwyg from '../../components/wysiwyg';
 
 import './style.scss';
 
 if (document.getElementById('post-create')) {
     new Vue({
         el: '#post-create',
-        components: {wysiwyg},
         mixins: [instance],
         render: h => h(postCreate),
     });
