+++ src/instances/post-create/style.scss
@@ -1,4 +1,5 @@
 @import '../../sass/organise';
+@import '~simplemde/dist/simplemde.min.css';
 
 #post-create-container {
     .submit-button {
